+++ core/src/main/scala-2.12/geotrellis/store/query/CirceCompat.scala
@@ -1,0 +1,3 @@
+package geotrellis.store.query
+
+object CirceCompat