+++ ogc-example/src/main/scala/geotrellis/server/ogc/conf/OgcServiceConf.scala
@@ -16,9 +16,8 @@
 
 package geotrellis.server.ogc.conf
 
-import geotrellis.server.ogc.ows
+import geotrellis.server.ogc.{OgcSourceRepository, SimpleSource, ows}
 import geotrellis.server.ogc.wms.WmsParentLayerMeta
-import geotrellis.server.ogc.{OgcSource, SimpleSource}
 import geotrellis.server.ogc.wmts.GeotrellisTileMatrixSet
 
 /**
@@ -28,12 +27,12 @@ import geotrellis.server.ogc.wmts.GeotrellisTileMatrixSet
  */
 sealed trait OgcServiceConf {
   def layerDefinitions: List[OgcSourceConf]
-  def layerSources(simpleSources: List[SimpleSource]): List[OgcSource] = {
+  def layerSources(simpleSources: List[SimpleSource]): OgcSourceRepository = {
     val simpleLayers =
-      layerDefinitions.collect { case ssc@SimpleSourceConf(_, _, _, _) => ssc.model }
+      layerDefinitions.collect { case ssc @ SimpleSourceConf(_, _, _, _) => ssc.models }.flatten
     val mapAlgebraLayers =
-      layerDefinitions.collect { case masc@MapAlgebraSourceConf(_, _, _, _) => masc.model(simpleSources) }
-    simpleLayers ++ mapAlgebraLayers
+      layerDefinitions.collect { case masc @ MapAlgebraSourceConf(_, _, _, _) => masc.model(simpleSources) }
+    OgcSourceRepository(simpleLayers ++ mapAlgebraLayers)
   }
 }